refactor(dataset): improve generating and loading sp500 data; docs(notebook): add short descriptions for generating sp500 components

+++ src/dataset.py
@@ -5,6 +5,7 @@ Code to download or generate data
 import shutil
 from datetime import datetime
 from pathlib import Path
+import pandas as pd
 
 import papermill as pm
 
@@ -12,9 +13,6 @@ from config import EXTERNAL_DATA_DIR, PROCESSED_DATA_DIR, PROJ_ROOT, RAW_DATA_DI
 
 
 class SP500:
-    """
-    Class to handle S&P 500 historical components
-    """
 
     submodule_name: str = "sp500"
     src_dir: Path = PROJ_ROOT / submodule_name
@@ -26,24 +24,49 @@ class SP500:
         pm.execute_notebook(str(nb_path), str(nb_path), cwd=str(nb_path.parent))
 
     @classmethod
-    def generate_current(cls):
-        nb_current: Path = cls.src_dir / "sp500.ipynb"
-        cls._run_notebook(nb_current)
-        ts = datetime.now().strftime("%Y-%m-%d")
-        dst_file: Path = cls.dst_dir / f"sp500_{ts}.csv"
-        shutil.copy(cls.src_dir / "sp500.csv", dst_file)
+    def _generate(cls, nb_path: Path, src_file: Path, dst_file: Path) -> Path:
+        if dst_file.exists():
+            print(f"No generation needed: {dst_file.name} already exists.")
+            return dst_file
+
+        cls._run_notebook(nb_path)
+        shutil.copy(src_file, dst_file)
+        print(f"Generated new file: {dst_file.name}")
         return dst_file
 
     @classmethod
-    def generate_historical(cls):
-        nb_historical: Path = cls.src_dir / "sp500_historical.ipynb"
-        cls._run_notebook(nb_historical)
-        ts_src = datetime.now().strftime("%m-%d-%Y")
+    def generate_current(cls) -> Path:
+        ts = datetime.now().strftime("%Y-%m-%d")
+        nb_current = cls.src_dir / "sp500.ipynb"
+        src_file = cls.src_dir / "sp500.csv"
+        dst_file = cls.dst_dir / f"sp500_{ts}.csv"
+        return cls._generate(nb_current, src_file, dst_file)
+
+    @classmethod
+    def generate_historical(cls) -> Path:
         ts_dst = datetime.now().strftime("%Y-%m-%d")
-        src_file: Path = cls.src_dir / f"S&P 500 Historical Components & Changes({ts_src}).csv"
-        dst_file: Path = cls.dst_dir / f"sp500_historical_{ts_dst}.csv"
-        shutil.copy(src_file, dst_file)
-        return dst_file
+        ts_src = datetime.now().strftime("%m-%d-%Y")
+        nb_historical = cls.src_dir / "sp500_historical.ipynb"
+        src_file = cls.src_dir / f"S&P 500 Historical Components & Changes({ts_src}).csv"
+        dst_file = cls.dst_dir / f"sp500_historical_{ts_dst}.csv"
+        return cls._generate(nb_historical, src_file, dst_file)
+
+    @classmethod
+    def _load_latest(cls, pattern: str) -> pd.DataFrame:
+        files = list(cls.dst_dir.glob(pattern))
+        if not files:
+            raise FileNotFoundError(f"No files for pattern: {pattern}")
+
+        latest_file = max(files, key=lambda f: f.name)
+        return pd.read_csv(latest_file)
+
+    @classmethod
+    def load_current(cls) -> pd.DataFrame:
+        return cls._load_latest("sp500_????-??-??.csv")
+
+    @classmethod
+    def load_historical(cls) -> pd.DataFrame:
+        return cls._load_latest("sp500_historical_????-??-??.csv")
 
 
 def main(
